FYAM-373: Survey - Updated value types for scale and range question types

+++ ForYouAndMe/Classes/Entities/Survey/SurveyQuestion.swift
@@ -30,9 +30,9 @@ struct SurveyQuestion {
     var image: UIImage?
     
     // Details
-    var minimum: Double?
-    var maximum: Double?
-    var interval: Double?
+    var minimum: Int?
+    var maximum: Int?
+    var interval: Int?
     @NilIfEmptyString
     var minimumDisplay: String?
     @NilIfEmptyString
